+++ R/export.R
@@ -330,7 +330,7 @@ exportSummary <- function(x, path = ".") {
       plotID <- names(modelPlots)[j]
       plotDisplay = modelPlots[[j]][["displayName"]]
       if (isEmpty(plotDisplay)) plotDisplay <- plotID
-      plotType <- modelPlots[[j]][["type"]]
+      plotType <- modelPlots[[j]][["plotType"]]
       if (isEmpty(plotType)) plotType <- "singleFeature"
       output[["plots"]][[i]][["plots"]][[j]] <- list(
         plotID = plotID,